+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [8.9.0] - 2020-04-30
+### Changed
+ - ProgramUtils and ProcessorUtils return dates instead of strings for date properties
+
 ## [8.8.1] - 2020-02-18
 ### Changed
  - IcdTimer - fixed issue that prevented OnElapsed event from firing when Length is less than (or close to) Heartbeat Interval